requirements.txt
asgiref==3.11.1
Django==6.0.4
mssql-django==1.7
pyodbc==5.3.0
python-dotenv==1.2.2
pytz==2026.1.post1
sqlparse==0.5.5
